Factored out ROGUE and BLEU metrics
