+++ flash/text/seq2seq/translation/model.py
@@ -17,7 +17,7 @@ import torch
 from torchmetrics import Metric
 
 from flash.text.seq2seq.core.model import Seq2SeqTask
-from flash.text.seq2seq.translation.metric import BLEUScore
+from flash.text.seq2seq.metrics import BLEUScore
 
 
 class TranslationTask(Seq2SeqTask):
